requirements.txt
flask
spacy
requests
python-dotenv